Separate Visualizers into Independent Classes
Vizualization used to be entangled with other methods
in Cammera and FerSenssor. Like emotion preddiction used to also draw
the bounding box around the face and text.

Now not only vizualization is not enangled with other methods, but also
all partial vizualizations are done in separate
PartialVizualizationCreator child classes, and all partial
vizaulizations are combined in VizualizaiotnAssembler.